requirements.txt
# Quant-coin-bot Dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
pyyaml>=6.0
pyjwt>=2.8.0
streamlit>=1.28.0
pytest>=7.0.0
pytest-cov>=4.0.0
